fs exfat: Disable exfat external module makefile logic
Our kernel build sometimes has the following build output:

    123148- MKELF scripts/mod/elfconfig.h
    123149- HOSTCC scripts/mod/modpost.o
    123150- HOSTCC scripts/dtc/data.o
    123151- HOSTLD scripts/genksyms/genksyms
    123152- HOSTCC scripts/selinux/genheaders/genheaders
    123153:scripts/basic/fixdep.c:462:1: fatal error: opening dependency
    file scripts/basic/.fixdep.d: Permission denied
    123154- }
    123155- ^
    123156-compilation terminated.
    123157- HOSTCC scripts/selinux/mdp/mdp
    123158- HOSTCC scripts/dtc/livetree.o
    123159- HOSTCC scripts/kallsyms
    123160:make[5]: *** [scripts/basic/fixdep] Error 1
    123161-make[4]: *** [scripts_basic] Error 2

The build succeeds, but we see the error.  This is due some odd
exfat makefile logic:

    ifneq ($(KERNELRELEASE),)
        # do the usual kernel module make
        ...

    else
        ...
        KDIR    := /lib/modules/$(shell uname -r)/build
        ...
        KREL   := $(shell cd ${KDIR} && make -s kernelrelease)
        ...

        all:
            $(MAKE) -C $(KDIR) M=$(PWD) modules
        ...
     endif

If KERNELRELEASE isn't set then KDIR points to the server's
Linux source.  In that case, we try to set KREL by
calling:

        cd <host kernel source>; make -s kernelrelease

This command, on my server, exits with the above error.  But
since the command is wrapped in a $(shell) it does not cause the build
to fail.

KERNELRELEASE is set via the main kernel Makefile:

        KERNELRELEASE = $(shell cat include/config/kernel.release 2>
                /dev/null)

During a parallel build the kernel.release file is sometimes not yet set
up when KERNELRELEASE is set (e.g., the first time kernel Makefile is
parsed during a clean build).  So it is not valid for a subordinate
makefile to use KERNELRELEASE for important build logic.

We don't ever want to build external kernel modules, so just disable
this faulty logic.

Change-Id: If1dbe3e54c41efb71e4c648fb45f46412f7c5abb
Signed-off-by: Andrew Wheeler <awheeler@motorola.com>
Reviewed-on: https://gerrit.mot.com/972154
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Shi-Yong Li <a22381@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Signed-off-by: SunnyRaj84348 <sunnyraj7461@gmail.com>